+++ src/components/check-box.js
@@ -41,7 +41,7 @@ export default React.createClass({
 
         this.setValue(value, () => {
             this.setDirtyState();
-            this.props.onChange(value, previousValue);
+            this.props.onChange(event, { value, previousValue });
         });
     },